refactor: engine 모듈 분리 및 이미지 토큰 공백 정규화
anchor-manager, diff-pipeline을 align-anchors, process-diff-elements로 분리.
build-diff-input에서 이미지 토큰 앞뒤 공백 정규화 추가. closes #78
