+++ core/src/diff/build-diff-input.ts
@@ -1,67 +1,4 @@
-// import { TOKEN_FLAGS_HAS_FOLLOWING_SPACE, TOKEN_FLAGS_LINE_END, TOKEN_FLAGS_LINE_START } from "../tokenization";
-// import type { DiffInput, DiffOptions } from "./types";
-
-// const DATA_STRIDE = 5;
-
-// export function buildDiffInput(wholeText: string, data: Int32Array, _diffOptions: DiffOptions): { input: DiffInput; lineCount: number } {
-//     const tokenCount = data.length / DATA_STRIDE;
-
-//     // 공백을 넣어줘야 자연스럽게 공백을 포함한 텍스트로 sa가 만들어짐.
-//     const insertSpace = _diffOptions.whitespace === "collapse";
-
-//     const flagsArray = new Uint32Array(tokenCount);
-//     const offsetArray = new Uint32Array(tokenCount + 1);
-
-//     let totalBufLen = 0;
-//     let lineCount = 0;
-//     for (let i = 0; i < tokenCount; i++) {
-//         const textLength = data[i * DATA_STRIDE + 1];
-//         const flags = data[i * DATA_STRIDE + 2];
-//         if (flags & TOKEN_FLAGS_LINE_START) {
-//             lineCount++;
-//         }
-
-//         flagsArray[i] = flags;
-//         offsetArray[i] = totalBufLen;
-
-//         totalBufLen += textLength;
-
-//         if (insertSpace && ((flags & TOKEN_FLAGS_HAS_FOLLOWING_SPACE) || (flags & TOKEN_FLAGS_LINE_END))) {
-//             totalBufLen++;
-//         }
-//     }
-//     offsetArray[tokenCount] = totalBufLen;
-
-//     const textBuffer = new Uint16Array(totalBufLen);
-
-//     let currentPos = 0;
-//     for (let i = 0; i < tokenCount; i++) {
-//         const ofs = data[i * DATA_STRIDE + 0];
-//         const len = data[i * DATA_STRIDE + 1];
-//         const flags = flagsArray[i];
-//         for (let j = 0; j < len; j++) {
-//             textBuffer[currentPos++] = wholeText.charCodeAt(ofs + j);
-//         }
-//         if (insertSpace && ((flags & TOKEN_FLAGS_HAS_FOLLOWING_SPACE) || (flags & TOKEN_FLAGS_LINE_END))) {
-//             textBuffer[currentPos++] = 32;
-//         }
-//     }
-
-//     data.fill(0);
-
-//     return {
-//         input: {
-//             buffer: textBuffer,
-//             offsets: offsetArray,
-//             flags: flagsArray,
-//             resultBuffer: data,
-//             tokenCount
-//         },
-//         lineCount
-//     };
-// }
-
-import { PAYLOAD_SHIFT, TOKEN_FLAGS_HAS_FOLLOWING_SPACE, TOKEN_FLAGS_LINE_END, TOKEN_FLAGS_LINE_START, TOKEN_FLAGS_TYPE_STRUCTURAL, TOKEN_TYPE_MASK } from "../tokenization";
+import { PAYLOAD_SHIFT, TOKEN_FLAGS_HAS_FOLLOWING_SPACE, TOKEN_FLAGS_LINE_END, TOKEN_FLAGS_LINE_START, TOKEN_FLAGS_TYPE_IMAGE, TOKEN_FLAGS_TYPE_STRUCTURAL, TOKEN_TYPE_MASK } from "../tokenization";
 import type { DiffInput, DiffOptions } from "./types";
 
 const DATA_STRIDE = 5;
@@ -90,6 +27,7 @@ export function buildDiffInput(wholeText: string, data: Int32Array, _diffOptions
 
     let totalBufLen = 0;
     let lineCount = 0;
+    let lastCharWasSpace = false;
     for (let i = 0; i < tokenCount; i++) {
         const textLength = data[i * DATA_STRIDE + 1];
         const flags = data[i * DATA_STRIDE + 2];
@@ -100,6 +38,13 @@ export function buildDiffInput(wholeText: string, data: Int32Array, _diffOptions
         }
 
         flagsArray[i] = flags;
+
+        // 이미지 토큰 앞에 공백 정규화
+        // 이미지 앞뒤로의 공백 여부는 whitespace 옵션과는 상관 없이 무시하는 것이 사람의 눈에 더 자연스러움!!
+        if (insertSpace && tokenType === TOKEN_FLAGS_TYPE_IMAGE && !lastCharWasSpace && totalBufLen > 0) {
+            totalBufLen++;
+        }
+
         offsetArray[i] = totalBufLen;
 
         if (tokenType === TOKEN_FLAGS_TYPE_STRUCTURAL) {
@@ -108,8 +53,18 @@ export function buildDiffInput(wholeText: string, data: Int32Array, _diffOptions
             totalBufLen += textLength;
         }
 
-        if (insertSpace && ((flags & TOKEN_FLAGS_HAS_FOLLOWING_SPACE) || (flags & TOKEN_FLAGS_LINE_END))) {
+        lastCharWasSpace = false;
+
+        // 이미지 토큰 뒤에 공백 정규화
+        // 이미지 앞뒤로의 공백 여부는 whitespace 옵션과는 상관 없이 무시하는 것이 사람의 눈에 더 자연스러움!!
+        if (insertSpace && tokenType === TOKEN_FLAGS_TYPE_IMAGE) {
             totalBufLen++;
+            lastCharWasSpace = true;
+        }
+
+        if (insertSpace && !lastCharWasSpace && ((flags & TOKEN_FLAGS_HAS_FOLLOWING_SPACE) || (flags & TOKEN_FLAGS_LINE_END))) {
+            totalBufLen++;
+            lastCharWasSpace = true;
         }
     }
     offsetArray[tokenCount] = totalBufLen;
@@ -117,11 +72,17 @@ export function buildDiffInput(wholeText: string, data: Int32Array, _diffOptions
     const textBuffer = new Uint16Array(totalBufLen);
 
     let currentPos = 0;
+    lastCharWasSpace = false;
     for (let i = 0; i < tokenCount; i++) {
         const ofs = data[i * DATA_STRIDE + 0];
         const len = data[i * DATA_STRIDE + 1];
         const flags = flagsArray[i];
         const tokenType = flags & TOKEN_TYPE_MASK;
+
+        // 이미지 토큰 앞에 공백 정규화
+        if (insertSpace && tokenType === TOKEN_FLAGS_TYPE_IMAGE && !lastCharWasSpace && currentPos > 0) {
+            textBuffer[currentPos++] = 32;
+        }
 
         if (tokenType === TOKEN_FLAGS_TYPE_STRUCTURAL) {
             if (STRUCTURAL_TOKEN_LENGTH > 0) {
@@ -137,8 +98,17 @@ export function buildDiffInput(wholeText: string, data: Int32Array, _diffOptions
             }
         }
 
-        if (insertSpace && ((flags & TOKEN_FLAGS_HAS_FOLLOWING_SPACE) || (flags & TOKEN_FLAGS_LINE_END))) {
+        lastCharWasSpace = false;
+
+        // 이미지 토큰 뒤에 공백 정규화
+        if (insertSpace && tokenType === TOKEN_FLAGS_TYPE_IMAGE) {
             textBuffer[currentPos++] = 32;
+            lastCharWasSpace = true;
+        }
+
+        if (insertSpace && !lastCharWasSpace && ((flags & TOKEN_FLAGS_HAS_FOLLOWING_SPACE) || (flags & TOKEN_FLAGS_LINE_END))) {
+            textBuffer[currentPos++] = 32;
+            lastCharWasSpace = true;
         }
     }
