+++ core/src/palette/default-palette.ts
@@ -7,7 +7,6 @@ export const DEFAULT_PALETTE: Readonly<Palette> = {
     diffAlpha: 1,
     diffLineColor: "hsl(0 100% 90% / 0.5)",
     highlightedDiffColor: "hsl(0 100% 80%)",
-    guidelineColor: "hsl(0 0% 50% / 0.3)",
     selectionHighlightColor: "hsl(0 0% 50% / 0.3)",
     minimapDiffColor: "hsl(0 100% 50% / 0.4)",
 };
