+++ core/src/palette/new-york-palette.ts
@@ -11,7 +11,6 @@ export const NEW_YORK_PALETTE: Readonly<Palette> = {
     diffAlpha: 1,
     diffLineColor: "hsl(220 13% 88% / 0.5)",
     highlightedDiffColor: "hsl(220 13% 75%)",
-    guidelineColor: "hsl(220 9% 46% / 0.2)",
     selectionHighlightColor: "hsl(220 9% 46% / 0.2)",
     minimapDiffColor: "hsl(220 18% 55% / 0.4)",
 };
