+++ core/src/renderer/editor-region.ts
@@ -3,7 +3,7 @@ import { Editor } from "../editor/editor";
 import { extractRectsFromRange } from "./extract-rects-from-range";
 import { mergeRects } from "./merge-rects";
 import { Renderer } from "./renderer";
-import { RENDER_FLAGS_DIFF_HIGHLIGHT, RENDER_FLAGS_DIFF_LAYER, RENDER_FLAGS_GEOMETRY, RENDER_FLAGS_HIGHLIGHT_LAYER, RENDER_FLAGS_MINIMAP, RENDER_FLAGS_RESIZE, RENDER_FLAGS_SCROLL, type DiffRenderItem, type Rect, type RectSet } from "./types";
+import { DIRTY_DIFF_HIGHLIGHT, DIRTY_GEOMETRY, DIRTY_RESIZE, DIRTY_SCROLL, DIRTY_SELECTION, RENDER_DIFF_LAYER, RENDER_HIGHLIGHT_LAYER, RENDER_MINIMAP, type DiffRenderItem, type Rect, type RectSet } from "./types";
 
 export class EditorRegion {
     readonly renderer: Renderer;
@@ -39,10 +39,10 @@ export class EditorRegion {
         let ret: number = 0;
         if (this.regionX !== x || this.regionY !== y || this.regionWidth !== width || this.regionHeight !== height) {
             renderer.invalidateGeometries(this.editor.name);
-            ret = RENDER_FLAGS_RESIZE;
+            ret = DIRTY_RESIZE;
         } else if (this.#scrollTop !== scrollTop) {
             renderer.invalidateScroll(this.editor.name);
-            ret = RENDER_FLAGS_SCROLL;
+            ret = DIRTY_SCROLL;
         }
 
         this.regionX = x;
@@ -101,8 +101,8 @@ export class EditorRegion {
         this.highlightedDiffIndex = diffIndex;
 
         if (wasShown || shouldShow) {
-            this.renderer.invalidateDiffLayer(this.editor.name);
-            return RENDER_FLAGS_DIFF_HIGHLIGHT;
+            this.renderer.invalidateRegion(DIRTY_DIFF_HIGHLIGHT, this.editor.name);
+            return DIRTY_DIFF_HIGHLIGHT;
         }
 
         return 0;
@@ -128,7 +128,7 @@ export class EditorRegion {
 
         this.#selectionHighlight = range;
         this.#selectionHighlightRects = null;
-        this.renderer.invalidateHighlightLayer(this.editor.name);
+        this.renderer.invalidateRegion(DIRTY_SELECTION, this.editor.name);
     }
 
     prepare(dirtyFlags: number) {
@@ -142,7 +142,7 @@ export class EditorRegion {
 
         visibleDiffIndices.clear();
 
-        if (dirtyFlags & RENDER_FLAGS_GEOMETRY) {
+        if (dirtyFlags & DIRTY_GEOMETRY) {
             diffGeometries.length = 0;
             this.#diffLineRects.length = 0;
             this.#sortedDiffIndices.length = 0;
@@ -264,11 +264,11 @@ export class EditorRegion {
     }
 
     render(dirtyFlags: number) {
-        if (dirtyFlags & RENDER_FLAGS_DIFF_LAYER) {
+        if (dirtyFlags & RENDER_DIFF_LAYER) {
             this.renderDiffLayer(dirtyFlags);
         }
 
-        if (dirtyFlags & RENDER_FLAGS_HIGHLIGHT_LAYER) {
+        if (dirtyFlags & RENDER_HIGHLIGHT_LAYER) {
             this.renderHighlightLayer(dirtyFlags);
         }
     }
@@ -278,7 +278,7 @@ export class EditorRegion {
         const ctx = this.renderer.ctx;
         ctx.save();
         ctx.translate(this.regionX, this.regionY);
-        if (dirtyFlags & RENDER_FLAGS_MINIMAP) {
+        if (dirtyFlags & RENDER_MINIMAP) {
             ctx.clearRect(0, 0, this.regionWidth, this.regionHeight);
         } else {
             ctx.clearRect(0, 0, this.regionWidth - this.renderer.options.minimapWidth, this.regionHeight);
@@ -333,7 +333,7 @@ export class EditorRegion {
             }
         }
 
-        if (dirtyFlags & RENDER_FLAGS_MINIMAP && this.diffGeometries.length > 0) {
+        if (dirtyFlags & RENDER_MINIMAP && this.diffGeometries.length > 0) {
             this.#renderMinimap(ctx);
         }
 
@@ -416,7 +416,7 @@ export class EditorRegion {
         const scrollLeft = 0;
 
         if (this.#selectionHighlight) {
-            if (!this.#selectionHighlightRects || dirtyFlags & RENDER_FLAGS_GEOMETRY) {
+            if (!this.#selectionHighlightRects || dirtyFlags & DIRTY_GEOMETRY) {
                 const offsetX = -this.renderer.x - this.regionX + scrollLeft;
                 const offsetY = -this.renderer.y - this.regionY + scrollTop;
                 const rawRects = extractRectsFromRange(this.#selectionHighlight, offsetX, offsetY, 0, 0, false);
